refactor quiz-app
