+++ react/src/machine-coding/quiz-app/types.ts
@@ -1,0 +1,5 @@
+export type ResultObj = {
+  score: number;
+  correctAnswers: number;
+  wrongAnswers: number;
+}
