+++ react/src/machine-coding/quiz-app/utils.ts
@@ -1,0 +1,1 @@
+export const addLeadingZero = (number: number) => (number > 9 ? number : `0${number}`);
